+++ src/gcp/cis-1.2.0/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@cloudgraph/policy-pack-gcp-cis-1.2.0-v1.23.1-alpha.1](https://github.com/cloudgraphdev/cloudgraph-policy-packs/compare/@cloudgraph/policy-pack-gcp-cis-1.2.0@1.23.0...@cloudgraph/policy-pack-gcp-cis-1.2.0@1.23.1-alpha.1) (2022-12-14)
+
+
+### Bug Fixes
+
+* **pnpm:** using semantic-release-pnpm ([eb9f3f2](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/eb9f3f22e85375b79be205c62adc09aa60628343))
+
 # [@cloudgraph/policy-pack-gcp-cis-1.2.0-v1.23.0](https://github.com/cloudgraphdev/cloudgraph-policy-packs/compare/@cloudgraph/policy-pack-gcp-cis-1.2.0@1.22.1...@cloudgraph/policy-pack-gcp-cis-1.2.0@1.23.0) (2022-08-01)