+++ src/gcp/cis-1.3.0/CHANGELOG.md
@@ -1,3 +1,13 @@
+# @cloudgraph/policy-pack-gcp-cis-1.3.0-v1.0.0-alpha.1 (2022-12-14)
+
+
+### Features
+
+* **CG-1291:** add gcp cis 130 boilerplate ([f7d6418](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/f7d6418826ba53ca588a552ad5a7b58300bfe468))
+* **CG-1292:** add GCP CIS 1.3 rules similar to 1.2 ([ff12a3f](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/ff12a3fed175cfeed414c4bf3d727ad65699230d))
+* **CG-1293:** add GCP CIS 1.30 1.16 rule ([286ea82](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/286ea828eaa472848613bd977663fd56002d1f78))
+* **CG-1293:** update the rule checker ([3e992cf](https://github.com/cloudgraphdev/cloudgraph-policy-packs/commit/3e992cf0d929c0e1e830e907e2ecc2a674cac77e))
+
 ### Features
 
 ### Bug Fixes